refactor board summary table header and body
